Adapted IPM book chapter 17 code to run with NIMBLE
The adaptation is the "bare minimum". I did not change the functions to simulate initial values, which are incomplete/too minimalist (NIMBLE is a bit more fussy than JAGS about missing/bad initial values and ideally, all nodes are given initial values). 

I also commented out some of the plotting code as the structure of the output from nimble and jagsUI is different and don't translate directly.